[MOV] bus: move timeout manager tests to dedicated file
The `TestWebsocketCaryall` test case includes a lot of unrelated
tests which makes the test intent hard to grasp. This commit moves
the timeout tests to a dedicated file.
